Better resign and beta estimate
1. Now we resign only if the condition is satisfied two moves in a row.

The resign condition is stricter when beta estimate is too small:

( if beta < 0.8 resign_thr /= 4 )

2. Much better beta estimate: if there are enough nodes, use the estimate of
the derivative of the mixture sigmoid at alpkt instead of the median.

(This typically gives slightly smaller values of beta, with exceptions.)

The new estimate is used in the output, in the training data, in the sgf
comments (before was raw net beta there) and in pass-agree conditions.

3. Improved output stats:

* 3 digits for beta (given also for every move)
* policy_loss is KL divergence instead of cross-entropy